requirements.txt
asgiref==3.3.3
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-environ==0.4.5
gunicorn==20.0.4
Pillow==8.1.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
